// backend/src/exames/enums/modalidade-dicom.enum.ts
export enum ModalidadeDicom {
  CR = 'CR', // Computed Radiography
  CT = 'CT', // Computed Tomography
  DX = 'DX', // Digital Radiography
  MG = 'MG', // Mammography
  MR = 'MR', // Magnetic Resonance
  NM = 'NM', // Nuclear Medicine
  OT = 'OT', // Other
  PT = 'PT', // Positron Emission Tomography
  RF = 'RF', // Radio Fluoroscopy
  US = 'US', // Ultrasound
  XA = 'XA', // X-Ray Angiography
}